proper handling of inf and nan in edge cases
inf and nan propagate naturally with the naive implementation, but
require special casing for correct results in the edge cases: having
both operands being inf/nan values and division/multiplication with 0.
the current implementation ensures that results are consistent with what
ieee-754 expects in the same situation, which "is derived from the
limiting cases of real arithmetic with operands of arbitrarily large
magnitude, when such a limit exists".