policy for switching between strict and permissive rationals
strict rationals do not allow 0 denominators. this allows for all the
special casing for inf/nan introduced in the last commit to be optimized
out, but it does introduce preconditions on construction and division.
permissive rationals do allow for 0 denominators, leading to inf/nan
special values. permissive rationals have no preconditions (except for
values which overflow) and all methods are noexcept.